feat: finish bot trade finish all function
Resolves:
See also:

+++ utils.py
@@ -15,13 +15,13 @@ def check_integer_format(text: str):
 def check_float_format(text: str):
     if text == "":
         return True
-    
+
     if all(x in "0123456789." for x in text) and text.count(".") <= 1:
         try:
-            int(text)
+            float(text)
             return True
         except ValueError:
             return False
-        
+
     else:
         return False
